robyol_lrp

+++ solo/methods/robyol_lrp.py
@@ -294,7 +294,7 @@ class RoBYOLLRP(BaseMomentumMethod):
             for v2 in np.delete(range(self.num_crops), v1):
 
                 # P = self.momentum_updater.cur_tau * apply_predictor(Z[v2], W) + (1-self.momentum_updater.cur_tau) * Z[v2]
-                W = closed_form_linear_predictor(Z[v1].detach(), Z_momentum[v2].detach())
+                W = closed_form_linear_predictor(Z[v1].detach().float(), Z_momentum[v2].detach().float())
                 P = apply_predictor(Z[v2], W)
                 # P = self.predictor(Z[v2])
                 neg_cos_sim += byol_loss_func(P, Z_momentum[v1])
